refactor: update import paths due to project structure changes
